Revert "Onboarding"
