+++ Shift/Features/Inspirations/InspirationDetailView.swift
@@ -18,12 +18,12 @@ struct InspirationDetailView: View {
         ScrollViewReader { proxy in
             ScrollView {
                 VStack(alignment: .leading) {
- //MARK: - Anchor to the top of the page for the navigation Buttons
+                    //MARK: Anchor to the top of the page for the navigation Buttons
                     Color.clear
                         .frame(height: 0)
                         .id("top")
                     
- //MARK: - Custom back button
+                    //MARK: Custom back button
                     Button(action: { dismiss() }) {
                         Text("< Timeline")
                             .font(.custom("HelveticaNeue", size: 17))
@@ -31,9 +31,9 @@ struct InspirationDetailView: View {
                     }
                     .padding(.top, 30)
                     .padding(.trailing)
-                
                     
- //MARK: - Image
+                    
+                    //MARK: Image
                     ZStack() {
                         Circle()
                             .fill(Color("NeonGreen"))
@@ -55,7 +55,7 @@ struct InspirationDetailView: View {
                     .frame(maxWidth: .infinity, maxHeight: .infinity)
                     .padding(.top)
                     
- //MARK: - Name + Field
+                    //MARK: Name + Field
                     Text(woman.name)
                         .font(.custom("Safiro-SemiBold", size: 34))
                         .foregroundColor(.white)
@@ -69,7 +69,7 @@ struct InspirationDetailView: View {
                         .frame(width: 372, height: 1)
                         .background(Color("NeonGreen"))
                     
- //MARK: - Birth – Death - Achievement
+                    //MARK: Birth – Death - Achievement
                     HStack{
                         Image("baby")
                             .renderingMode(.template)
@@ -117,7 +117,8 @@ struct InspirationDetailView: View {
                         .foregroundColor(.clear)
                         .frame(width: 372, height: 1)
                         .background(Color("NeonGreen"))
- //MARK: - Biography
+                    
+                    //MARK: Biography
                     Text("Biography")
                         .font(.custom("Safiro-SemiBold", size: 24))
                         .foregroundColor(.white)
@@ -128,7 +129,7 @@ struct InspirationDetailView: View {
                         .font(.custom("HelveticaNeue", size: 14))
                         .foregroundColor(.white)
                     
- //MARK: - Why Inspiring
+                    //MARK: Why Inspiring
                     Text("Why She's Inspiring")
                         .font(.custom("Safiro-SemiBold", size: 24))
                         .foregroundColor(.white)
@@ -139,7 +140,7 @@ struct InspirationDetailView: View {
                         .font(.custom("HelveticaNeue", size: 14))
                         .foregroundColor(.white)
                     
-                    // - MARK: Navigation Buttons
+                    // MARK: Navigation Buttons
                     HStack {
                         Button(action: {
                             if currentIndex > 0 {
@@ -157,7 +158,6 @@ struct InspirationDetailView: View {
                                 .cornerRadius(12)
                         }
                         .disabled(currentIndex == 0)
-                        .opacity(currentIndex == 0 ? 0.7 : 1.0)
                         
                         Spacer()
                         
@@ -177,16 +177,13 @@ struct InspirationDetailView: View {
                                 .cornerRadius(12)
                         }
                         .disabled(currentIndex == women.count - 1)
-                        .opacity(currentIndex == women.count - 1 ? 0.7 : 1.0)
                     }
                     .padding(.top, 35)
                     .padding(.bottom, 32)
                     .font(.custom("Safiro-SemiBold", size: 16))
                 }
                 .padding()
-                .padding(.bottom, 55)
             }
-            
             .background(Color("Violet"))
             .edgesIgnoringSafeArea(.all)
             .navigationBarBackButtonHidden(true)
