+++ Shift/Features/Inspirations/InspirationsModel.swift
@@ -15,7 +15,7 @@ struct Inspirations: Identifiable, Codable { // Identifiable -  uses the id prop
     let imageWhite: String
     let imageGreen: String
     let birth: String
-    let death: String? 
+    let death: String?      // Optional – can be nil if still alive
     let achievement: String
     let biography: String
     let whyInspiring: String
