+++ Shift/Features/Shared/Models/User.swift
@@ -23,5 +23,3 @@ struct User: Identifiable {
         self.interests = interests
     }
 }
-//
-
