+++ Shift/Resources/Local Data/inspirational_women.json
@@ -113,7 +113,7 @@
         "field": "Technology Executive",
         "imageWhite": "ginni_rometty",
         "imageGreen": "ginni_rometty_g",
-        "birth": "July 29, 1957, USA",
+        "birth": "July 29, 1957, Chicago, Illinois",
         "death": null,
         "achievement": "First female CEO of IBM; led company through cloud and AI transformation",
         "biography": "Virginia 'Ginni' Rometty was the first woman to serve as CEO of IBM, a position she held from 2012 to 2020. Under her leadership, IBM shifted focus toward cloud computing, artificial intelligence, and data analytics, transforming its business model for the digital era. Rometty championed innovation, workforce diversity, and strategic acquisitions that positioned IBM as a leader in emerging technologies. She is recognized for her commitment to STEM education and expanding opportunities for women and minorities in tech.",
